for the number of clusters uint32 is enough

+++ src/public/utils.hpp
@@ -57,7 +57,7 @@ MPI_Datatype createLocalCentroidDataType(MPI_Datatype pointType) {
     return localCentroidType;
 };
 
-void writeResultsToFile(const std::vector<uint64>& memberships,
+void writeResultsToFile(const std::vector<uint32>& memberships,
                         const char *in = "../data/in.csv",
                         const char *out = "../data/out.csv") {
     FILE *fp = fopen(in, "r");
@@ -69,7 +69,7 @@ void writeResultsToFile(const std::vector<uint64>& memberships,
     float32 x, y;
     uint64 processed = 0;
     while (fscanf(fp, "%f,%f,%*f", &x, &y) > 0) {
-        fprintf(fp_out, "%f,%f,%llu\n", x, y, memberships[processed]);
+        fprintf(fp_out, "%f,%f,%u\n", x, y, memberships[processed]);
         processed++;
     }
